update 27/1 1:12

+++ BE/Configuration.cs
@@ -12,7 +12,7 @@ namespace BE
         public static long HostID = 0;
         public static long HostingUnitID = 0;
         public static long OrderID = 0;
-        public static int commission = 10;
+        public static float commission = 10;
         public static int orderValidTime = 30;
         public static int orderCancelationDays = 14;
     }
